Move app related actions into appActions.js
